+++ src/CloseButton.js
@@ -1,8 +1,6 @@
 import * as PropTypes from 'prop-types';
 import React from 'react';
 
-import Button from './Button';
-
 const propTypes = {
   label: PropTypes.string.isRequired,
   onClick: PropTypes.func,
@@ -12,13 +10,14 @@ class CloseButton extends React.Component {
   render() {
     const { label, onClick } = this.props;
     return (
-      <Button
+      <button
+        type="button"
         className="close"
         onClick={onClick}
       >
         <span aria-hidden="true">&times;</span>
         <span className="sr-only">{label}</span>
-      </Button>
+      </button>
     );
   }
 }
