+++ test/PanelGroupSpec.js
@@ -1,6 +1,8 @@
 import React from 'react';
 import ReactTestUtils from 'react-addons-test-utils';
+import ReactDOM from 'react-dom';
 
+import Accordion from '../src/Accordion';
 import Panel from '../src/Panel';
 import PanelGroup from '../src/PanelGroup';
 
@@ -88,6 +90,21 @@ describe('<PanelGroup>', () => {
     );
 
     assert.ok(panel.state.expanded);
+  });
+
+  describe('<Accordion>', () => {
+    it('Should render a div with a role of tablist', () => {
+      let instance = ReactTestUtils.renderIntoDocument(
+        <Accordion>
+          Accordion content
+        </Accordion>
+      );
+
+      const accordion = ReactDOM.findDOMNode(instance);
+
+      assert.equal(accordion.nodeName, 'DIV');
+      assert.equal(accordion.getAttribute('role'), 'tablist');
+    });
   });
 
   describe('Web Accessibility', () => {
